+++ record/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 6.2.1
 * fix: Stop requesting for amplitude updates when recording is paused or stopped from native side.
 * chore: Code cleanup.
 * chore: Remove uuid package dependency.